Add default constructor to ClapTrap class
This commit introduces a default constructor to the ClapTrap class. Previously, instances of this class could only be created with a specified name. Now, a ClapTrap with a 'default' name and initial health, energy and attack values can be generated if no name is provided. This allows more flexibility in creating ClapTrap instances. Also, output messaging has been made more specific to the type of constructor used.