Add ScavTrap default constructor
This commit adds a default constructor to the ScavTrap class. It inherits attributes from the ClapTrap class. This allows creating a ScavTrap with default name, health, energy, and attack values if no name is provided. A default constructor provides more flexibility to create instances of the ScavTrap class. The output messages have also been updated to correctly represent the constructor used.