Add DiamondTrap and update files
This commit introduces the DiamondTrap class which extends the inherited classes FragTrap and ScavTrap. A field "_name" and a new constructor for initializing "name" were added for ScavTrap and FragTrap. Files which needed to link with these new classes were updated including related header and .cpp files. The changes were necessary for adding the specific functionality of the DiamondTrap. Along with this, .gitignore file has been added in the repository to avoid keeping unnecessary and subsidiary files. The code lines in other classes were also modified for better output representation.